+++ src/lib/style/Portal.css
@@ -392,13 +392,11 @@ table.SRC-paddingRight > tbody > tr > td {
 
 .SRC-primary-background-color-hover:hover {
   background-color: #D53F77; }
-
-.SRC-primary-background-color-hover:hover > svg {
-  color: white; }
-
-.SRC-primary-background-color-hover:hover.SRC-nested-color:hover * {
-  background-color: #D53F77;
-  color: white; }
+  .SRC-primary-background-color-hover:hover > svg {
+    color: white; }
+  .SRC-primary-background-color-hover:hover.SRC-nested-color:hover * {
+    background-color: #D53F77;
+    color: white; }
 
 .SRC-marginFifteen {
   margin-left: 15px;
@@ -549,9 +547,8 @@ button.SRC-roundBorder {
   font-weight: 600;
   border: none;
   background: white; }
-
-.tableDropdownSelector:focus {
-  outline: none; }
+  .tableDropdownSelector:focus {
+    outline: none; }
 
 .SRC-extraPadding {
   padding: 5px; }
@@ -577,9 +574,8 @@ button.SRC-roundBorder {
   width: 130px;
   height: 40px;
   margin: 21px; }
-
-.SRC-viewMoreButton:focus {
-  outline: none; }
+  .SRC-viewMoreButton:focus {
+    outline: none; }
 
 /*
   checkmark styling start
@@ -674,9 +670,8 @@ checkmark styling end
   z-index: 900;
   background: none;
   border: none; }
-
-.SRC-menu-wall:focus {
-  outline: none; }
+  .SRC-menu-wall:focus {
+    outline: none; }
 
 .SRC-table-dropdown-content {
   align-items: center;
@@ -815,10 +810,9 @@ div.SRC-userImg {
   display: inline-flex;
   flex-direction: column;
   padding-left: 17px; }
-
-.SRC-cardContent p {
-  margin: 0px 0px 4px 0px;
-  line-height: 120%; }
+  .SRC-cardContent p {
+    margin: 0px 0px 4px 0px;
+    line-height: 120%; }
 
 .SRC-textareaCopy {
   border: none;
