+++ src/lib/utils/SynapseClient.tsx
@@ -13,6 +13,11 @@ export const DEV_ENV_SESSION_LOCAL_STORAGE_KEY = 'session-token-dev-mode-only'
 const DEFAULT_ENDPOINT = 'https://repo-prod.prod.sagebase.org/'
 const DEFAULT_SWC_ENDPOINT = 'https://www.synapse.org/'
 
+export const AUTH_PROVIDER = 'GOOGLE_OAUTH_2_0'
+export const getRootURL = () => {
+  return `${window.location.protocol}//${window.location.hostname}:${window.location.port}/`
+}
+
 function delay(t: any) {
   return new Promise((resolve) => {
     setTimeout(resolve.bind(null, null), t)
@@ -517,4 +522,53 @@ export const getPrincipalAliasRequest = (sessionToken: string | undefined,
                                          endpoint: string = DEFAULT_ENDPOINT) => {
   const url = 'repo/v1/principal/alias'
   return doPost(url, { alias, type }, sessionToken, undefined, endpoint)
+}
+
+/*
+During SSO login, the authorization provider (Google) will
+send the user back to the portal with an authorization code,
+which can be exchanged for a Synapse user session.
+This function should be called whenever the root App is initialized
+(to look for this code parameter and complete the round-trip).
+*/
+export const detectSSOCode = () => {
+  const redirectURL = getRootURL()
+  // 'code' handling (from SSO) should be preformed on the root page, and then redirect to original route.
+  let code: URL | null | string = new URL(window.location.href)
+  // in test environment the searchParams isn't defined
+  const { searchParams } = code
+  if (!searchParams) {
+    return
+  }
+  code = searchParams.get('code')
+  if (code) {
+    oAuthSessionRequest(AUTH_PROVIDER, code, `${redirectURL}?provider=${AUTH_PROVIDER}`)
+              .then((synToken: any) => {
+                setSessionTokenCookie(synToken.sessionToken).then(() => {
+                  // go back to original route after successful SSO login
+                  const originalUrl = localStorage.getItem('after-sso-login-url')
+                  localStorage.removeItem('after-sso-login-url')
+                  if (originalUrl) {
+                    window.location.replace(originalUrl)
+                  }
+                }).catch((errSetSession) => {
+                  console.error('Error on set sesion token cookie ', errSetSession)
+                })
+              })
+              .catch((err: any) => {
+                if (err.statusCode === 404) {
+                  // Synapse account not found, send to registration page
+                  window.location.replace('https://www.synapse.org/#!RegisterAccount:0')
+                }
+                console.error('Error on sso sign in ', err)
+              })
+  }
+}
+
+export const signOut = () => {
+  setSessionTokenCookie(undefined).then(() => {
+    window.location.reload()
+  }).catch((err) => {
+    console.error('err when clearing the session cookie ', err)
+  })
 }
